Move tutorials to a separate package
