+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/tutorials/TutorialControllerAndSensorsTeam2.java
@@ -1,6 +1,6 @@
 // Date: 7-23-24
 
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.tutorials;
 
 import com.qualcomm.robotcore.eventloop.opmode.OpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
@@ -9,17 +9,17 @@ import com.qualcomm.robotcore.hardware.DcMotor;
 import com.qualcomm.robotcore.hardware.TouchSensor;
 
 @TeleOp(name="Tutorial: Controller and Sensors Turns Motor")
-public class TutorialControllerAndSensorsTeam1 extends OpMode {
+public class TutorialControllerAndSensorsTeam2 extends OpMode {
     DcMotor motor;
-    TouchSensor touchSensor;
-    ColorSensor jimmy;
+    TouchSensor samuel;
+    ColorSensor bob;
     double power;
 
     @Override
     public void init() {
         motor = hardwareMap.get(DcMotor.class, "motory");
-        jimmy = hardwareMap.get(ColorSensor.class, "colory");
-        touchSensor = hardwareMap.get(TouchSensor.class, "touchy");
+        bob = hardwareMap.get(ColorSensor.class, "colory");
+        samuel = hardwareMap.get(TouchSensor.class, "touchy");
     }
 
     @Override
@@ -30,19 +30,19 @@ public class TutorialControllerAndSensorsTeam1 extends OpMode {
 
         // If no triggers pressed, use color sensor. Green = forward. Red = backward.
         if ((gamepad1.right_trigger == 0) && (gamepad1.left_trigger == 0)) {
-            int bob = jimmy.red();
-            int tim = jimmy.green();
-            power = (double)(tim - bob) / 255;
+            int bred = bob.red();
+            int bgreen = bob.green();
+            power = (double)(bgreen - bred) / 10000;
         }
 
         // If touch sensor pressed, kill switch.
-        if(touchSensor.isPressed()) {
+        if(samuel.isPressed()) {
             power = 0;
         }
 
-        telemetry.addData("Red", jimmy.red());
-        telemetry.addData("Green", jimmy.green());
-        telemetry.addData("ButtonPressed", touchSensor.isPressed());
+        telemetry.addData("Red", bob.red());
+        telemetry.addData("Green", bob.green());
+        telemetry.addData("ButtonPressed", samuel.isPressed());
         telemetry.addData("Power", power);
         telemetry.update();
         motor.setPower(power);
